+++ blockly/src/entities/monster/StraightRangeAI.java
@@ -1,13 +1,13 @@
 package entities.monster;
 
+import coderunner.BlocklyCommands;
 import contrib.utils.components.skill.ISkillUser;
 import contrib.utils.components.skill.Skill;
 import core.Entity;
 import core.Game;
 import core.utils.MissingHeroException;
-import java.util.function.Consumer;
-import utils.BlocklyCommands;
 import entities.EntityUtils;
+import java.util.function.Consumer;
 
 /**
  * Attacks the player if he is the view range of the entity. The entity will only shoot in its view
