+++ blockly/src/level/produs/Level009.java
@@ -6,8 +6,8 @@ import core.level.utils.Coordinate;
 import core.level.utils.DesignLabel;
 import core.level.utils.LevelElement;
 import core.utils.Direction;
-import entities.monster.BlocklyMonster;
 import entities.MiscFactory;
+import entities.monster.BlocklyMonster;
 import java.util.List;
 import level.BlocklyLevel;
 import level.LevelManagementUtils;
