+++ blockly/src/level/produs/Level011.java
@@ -12,8 +12,8 @@ import core.level.utils.DesignLabel;
 import core.level.utils.LevelElement;
 import core.utils.Direction;
 import core.utils.components.MissingComponentException;
-import entities.monster.BlocklyMonster;
 import entities.MiscFactory;
+import entities.monster.BlocklyMonster;
 import java.util.List;
 import level.BlocklyLevel;
 import level.LevelManagementUtils;
