+++ blockly/src/level/produs/Level021.java
@@ -13,8 +13,8 @@ import core.level.utils.LevelElement;
 import core.utils.Direction;
 import core.utils.MissingHeroException;
 import core.utils.components.MissingComponentException;
-import entities.monster.BlocklyMonster;
 import entities.MiscFactory;
+import entities.monster.BlocklyMonster;
 import java.util.List;
 import level.BlocklyLevel;
 import level.LevelManagementUtils;
